pyproject.toml
[project]
name = "fomc-hawkish-dovish"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "datasets>=3.0.2",
    "flair>=0.14.0",
    "hydra-core>=1.3.2",
    "ipykernel>=6.29.5",
    "ipywidgets>=8.1.5",
    "kaleido==0.2.1",
    "lightning>=2.4.0",
    "optuna-dashboard>=0.16.2",
    "optuna>=4.0.0",
    "plotly>=5.24.1",
    "polars[gpu]>=1.11.0",
    "sentence-transformers>=3.2.1",
    "torch>=2.5.0",
    "torchmetrics>=1.5.1",
    "transformers>=4.46.0",
    "umap-learn>=0.5.6",
    "wandb>=0.18.5",
]
